+++ Runtime/Jobs/UpdateJob.cs
@@ -2,9 +2,6 @@ using Gilzoide.UpdateManager.Jobs.Internal;
 
 namespace Gilzoide.UpdateManager.Jobs
 {
-#if HAVE_BURST
-    [Unity.Burst.BurstCompile]
-#endif
     public struct UpdateJob<TData> : IInternalUpdateJob<TData>
         where TData : struct, IUpdateJob
     {
@@ -15,4 +12,18 @@ namespace Gilzoide.UpdateManager.Jobs
             Data.ItemRefAt(index).Execute();
         }
     }
+
+#if HAVE_BURST
+    [Unity.Burst.BurstCompile]
+    public struct BurstUpdateJob<TData> : IInternalUpdateJob<TData>
+        where TData : struct, IUpdateJob
+    {
+        public UnsafeNativeList<TData> Data { get; set; }
+
+        public unsafe void Execute(int index)
+        {
+            Data.ItemRefAt(index).Execute();
+        }
+    }
+#endif
 }