+++ Runtime/Jobs/UpdateTransformJob.cs
@@ -3,9 +3,6 @@ using Gilzoide.UpdateManager.Jobs.Internal;
 
 namespace Gilzoide.UpdateManager.Jobs
 {
-#if HAVE_BURST
-    [Unity.Burst.BurstCompile]
-#endif
     public struct UpdateTransformJob<TData> : IInternalUpdateTransformJob<TData>
         where TData : struct, IUpdateTransformJob
     {
@@ -16,4 +13,18 @@ namespace Gilzoide.UpdateManager.Jobs
             Data.ItemRefAt(index).Execute(transform);
         }
     }
+
+#if HAVE_BURST
+    [Unity.Burst.BurstCompile]
+    public struct BurstUpdateTransformJob<TData> : IInternalUpdateTransformJob<TData>
+        where TData : struct, IUpdateTransformJob
+    {
+        public UnsafeNativeList<TData> Data { get; set; }
+
+        public unsafe void Execute(int index, TransformAccess transform)
+        {
+            Data.ItemRefAt(index).Execute(transform);
+        }
+    }
+#endif
 }