+++ Samples~/FollowTarget/FollowTarget.cs
@@ -23,7 +23,11 @@ namespace Gilzoide.UpdateManager.Sample.FollowTarget
         }
     }
 
-    public class FollowTarget : AJobBehaviour<FollowTargetJob, UpdateTransformJob<FollowTargetJob>>
+#if HAVE_BURST
+    public class FollowTarget : AJobBehaviour<FollowTargetJob, BurstUpdateTransformJob<FollowTargetJob>>
+#else
+    public class FollowTarget : AJobBehaviour<FollowTargetJob>
+#endif
     {
         [SerializeField] private Target target;
         [SerializeField] private float speed;