+++ Samples~/FollowTarget/Target.cs
@@ -16,7 +16,11 @@ namespace Gilzoide.UpdateManager.Sample.FollowTarget
         }
     }
 
-    public class Target : AJobBehaviour<TargetJob, UpdateTransformJob<TargetJob>>
+#if HAVE_BURST
+    public class Target : AJobBehaviour<TargetJob, BurstUpdateTransformJob<TargetJob>>
+#else
+    public class Target : AJobBehaviour<TargetJob>
+#endif
     {
         public NativeReference<Vector3> PositionReference => positionReference;
         private NativeReference<Vector3> positionReference;